Modified migrations, control to store pictures in sections and dishes added, modified exceptions

+++ database/migrations/2021_03_24_201531_create_dishes_table.php
@@ -13,21 +13,21 @@ class CreateDishesTable extends Migration
      */
     public function up()
     {
-        Schema::create('dishes', function (Blueprint $table) {
-            $table->id();
-            $table->string('name', 255);
-            $table->float('price');
-            $table->integer('units')->nullable();
-            $table->boolean('extra')->default(false);
-            $table->boolean('hidden')->default(false);
-            $table->boolean('menu')->default(false);
-            $table->float('price_menu')->nullable();
-            $table->text('ingredients')->nullable();
-            $table->foreignId('section_id')->constrained()->onDelete('cascade');
-            $table->foreignId('picture_id')->constrained();
-
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('dishes')) {
+            Schema::create('dishes', function (Blueprint $table) {
+                $table->id();
+                $table->string('name', 255);
+                $table->float('price');
+                $table->integer('units')->nullable();
+                $table->boolean('extra')->default(false);
+                $table->boolean('hidden')->default(false);
+                $table->boolean('menu')->default(false);
+                $table->float('price_menu')->nullable();
+                $table->text('ingredients')->nullable();
+                $table->foreignId('section_id')->constrained()->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
